+++ package.json
@@ -1,10 +1,13 @@
 {
   "dependencies": {
+    "@react-oauth/google": "^0.12.1",
     "all": "^0.0.0",
     "js-cookie": "^3.0.5",
     "node-sass": "^9.0.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-google-button": "^0.8.0",
+    "react-scripts": "^5.0.1",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
